loadtest/requirements.txt
playwright>=1.53.0
rich>=13.7.0
pyyaml>=6.0.1
httpx>=0.27.0
